refactor: split ai chat and public home state
