+++ frontend/src/features/lms/pages/PublicHomePage.tsx
@@ -1,6 +1,6 @@
-import { useMemo, useRef, useState } from 'react';
 import type { CourseCard } from '@myway/shared';
 import { PublicHomeCourseSection, PublicHomeHeroSection } from './PublicHomePageSections';
+import { usePublicHomePageState } from './usePublicHomePageState';
 
 type PublicHomePageProps = {
   courseCards: CourseCard[];
@@ -16,64 +16,25 @@ const navItems = [
   { label: '커뮤니티', icon: 'ri-group-line' },
 ];
 
-function countUnique(values: string[]): number {
-  return new Set(values).size;
-}
-
-function scrollToElement(element: HTMLElement | null) {
-  element?.scrollIntoView({ behavior: 'smooth', block: 'start' });
-}
-
 export function PublicHomePage({ courseCards, busy, onOpenLogin }: PublicHomePageProps) {
-  const coursesRef = useRef<HTMLElement | null>(null);
-  const roadmapRef = useRef<HTMLElement | null>(null);
-  const [searchQuery, setSearchQuery] = useState('');
-  const [activeCategory, setActiveCategory] = useState('');
-  const [activeStatus, setActiveStatus] = useState<'all' | 'available' | 'enrolled'>('all');
-
-  const categories = useMemo(() => [...new Set(courseCards.map((item) => item.category))], [courseCards]);
-
-  const filteredCourses = useMemo(() => {
-    const query = searchQuery.trim().toLowerCase();
-
-    return courseCards.filter((course) => {
-      const tags = Array.isArray(course.tags) ? course.tags : [];
-      const queryMatch = query
-        ? [course.title, course.description, course.category, course.instructor_name, ...tags].join(' ').toLowerCase().includes(query)
-        : true;
-      const categoryMatch = activeCategory ? course.category === activeCategory : true;
-      const statusMatch =
-        activeStatus === 'all'
-          ? true
-          : activeStatus === 'available'
-            ? !course.enrolled
-            : course.enrolled;
-
-      return queryMatch && categoryMatch && statusMatch;
-    });
-  }, [activeCategory, activeStatus, courseCards, searchQuery]);
-
-  const featuredCourses = filteredCourses.slice(0, 5);
-  const featuredTags = useMemo(() => {
-    const counts = new Map<string, number>();
-
-    courseCards.forEach((course) => {
-      const tags = Array.isArray(course.tags) ? course.tags : [];
-      tags.forEach((tag) => {
-        counts.set(tag, (counts.get(tag) ?? 0) + 1);
-      });
-    });
-
-    return [...counts.entries()]
-      .sort((left, right) => right[1] - left[1] || left[0].localeCompare(right[0]))
-      .slice(0, 6)
-      .map(([tag]) => tag);
-  }, [courseCards]);
-
-  const totalProgress = Math.round(
-    courseCards.reduce((sum, course) => sum + course.progress_percent, 0) / Math.max(courseCards.length, 1),
-  );
-  const tagCount = countUnique(courseCards.flatMap((course) => (Array.isArray(course.tags) ? course.tags : [])));
+  const {
+    coursesRef,
+    roadmapRef,
+    searchQuery,
+    setSearchQuery,
+    activeCategory,
+    setActiveCategory,
+    activeStatus,
+    setActiveStatus,
+    categories,
+    filteredCourses,
+    featuredCourses,
+    featuredTags,
+    totalProgress,
+    tagCount,
+    scrollToCourses,
+    scrollToRoadmap,
+  } = usePublicHomePageState({ courseCards });
 
   return (
     <div className="min-h-screen bg-[#eef2f7]">
@@ -92,12 +53,12 @@ export function PublicHomePage({ courseCards, busy, onOpenLogin }: PublicHomePag
                 key={item.label}
                 type="button"
                 onClick={() => {
-                  if (item.label === '강의 탐색') {
-                    scrollToElement(coursesRef.current);
+              if (item.label === '강의 탐색') {
+                    scrollToCourses();
                     return;
                   }
                   if (item.label === '학습 로드맵') {
-                    scrollToElement(roadmapRef.current);
+                    scrollToRoadmap();
                     return;
                   }
                   if (item.label === 'AI 도구' || item.label === '커뮤니티') {
